requirements_test.txt
pytest==7.4.0
pytest-homeassistant-custom-component==0.13.41
pytest-asyncio==0.21.1
aioresponses==0.7.4